Cleanup container cli and methods
