+++ janisdk/container/__init__.py
@@ -1,38 +1,17 @@
-import argparse, sys, os, subprocess
-
-from janis_core import Logger
-
-from toolbuilder.parse_help import from_container
-from toolbuilder.templates import ToolTemplateType
-from toolbuilder.runtest import runner as test_runner
-
-from janis_assistant.management.configuration import JanisConfiguration
-
-
-def process_args():
-    cmds = {"container": do_container, "run-test": do_runtest}
-
-    parser = argparse.ArgumentParser(description="Execute a workflow")
-    subparsers = parser.add_subparsers(help="subcommand help", dest="command")
-    parser.add_argument("-d", "--debug", action="store_true")
-
-    subparsers.add_parser("version")
-    add_container_args(subparsers.add_parser("container"))
-    test_runner.add_runtest_args(subparsers.add_parser("run-test"))
-    # add_workflow_args(subparsers.add_parser("run-workflow"))
-
-    args = parser.parse_args()
-    return cmds[args.command](args)
+import sys
 
 
 def do_container(args):
+    from janis_core.translations.janis import ToolTemplateType
+    from janisdk.container.parse_help import from_container
+
     tooltype = ToolTemplateType.base
     if args.gatk4:
         tooltype = ToolTemplateType.gatk4
 
     outputdir = args.output
 
-    (tool, toolversion), helpstr = from_container(
+    tool, helpstr = from_container(
         container=args.container,
         basecommand=args.basecommand,
         helpcommand=args.help_str,
@@ -56,13 +35,13 @@ def do_container(args):
 
         with open(path.join(outputdir, "base.py"), "w+") as f:
             f.write(tool)
-        with open(path.join(outputdir, "versions.py"), "w+") as f:
-            f.write(toolversion)
+        # with open(path.join(outputdir, "versions.py"), "w+") as f:
+        #     f.write(toolversion)
         with open(path.join(outputdir, "__init__.py"), "w+"):
             pass
     else:
         print(tool, file=sys.stdout)
-        print(toolversion, file=sys.stdout)
+        # print(toolversion, file=sys.stdout)
 
 
 def add_container_args(parser):
@@ -114,27 +93,3 @@ def add_container_args(parser):
     extra.add_argument(
         "--gatk4", action="store_true", help="Use the GATK4 tool template"
     )
-
-
-def do_runtest(args):
-    config = None
-    if args.config:
-        config = JanisConfiguration.initial_configuration(path=args.config)
-
-    runner_path = test_runner.__file__
-
-    cli_args = sys.argv[2:]
-    run_test_commands = ["python", runner_path] + cli_args
-
-    if config:
-        commands = config.template.template.prepare_run_test_command(run_test_commands)
-    else:
-        commands = run_test_commands
-
-    joined_command = "' '".join(commands)
-    Logger.info(f"Deploying test with command: '{joined_command}'")
-    subprocess.run(commands)
-
-
-if __name__ == "__main__":
-    process_args()
