--- toolbuilder/parse_cwl.py
@@ -1,17 +1,0 @@
-import cwlgen
-
-
-def from_cwlgen():
-    pass
-
-
-def from_dict():
-    pass
-
-
-def from_yamlstr():
-    pass
-
-
-def from_file():
-    pass
